requirements.txt
PyQt6
requests
python-vlc
eyed3